last update 22 mordad 1401

+++ lib/Models/Home/pricey_courses_model.dart
@@ -36,7 +36,7 @@ class PriceyCoursesModel {
   String? views;
   List<int>? update;
   int? reviews;
-  int? reviewsRating;
+  double? reviewsRating;
 
   RxBool visible;
